pyproject.toml
[project]
name = "langchain-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.17",
    "langchain[openai]>=0.3.25",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
]
